+++ src/Entities/Ghost.cpp
@@ -8,7 +8,7 @@
 
 using namespace std;
 
-Ghost::Ghost(int posX, int posY,int speed) : MovingEntity(posX, posY,speed){
+Ghost::Ghost(float posX, float posY,float speed) : MovingEntity(posX, posY,speed){
     this->type = GHOST;
     this->STATE = DIR_DOWN;
 }
